Don't post distributed notifications in a mac sandboxed app
